FAC: Move trace logic to execution module
